requirements.txt
openai>=1.43.0
tiktoken>=0.7.0

langchain>=0.3.2
langchain-community>=0.3.1
langchain-openai>=0.1.24
chromadb>=0.5.4
pypdf>=5.0.0

# RAGAS and deps
ragas>=0.1.9
datasets>=2.20.0
evaluate>=0.4.2
pandas>=2.2.2
numpy>=2.1.2
openpyxl>=3.1.0

python-dotenv>=1.0.1

# Web interface
streamlit>=1.28.0

# Advanced RAG features
beautifulsoup4>=4.12.0
lxml>=5.0.0
rank-bm25>=0.2.2
sentence-transformers>=2.2.2

# Alternative embedding models (optional)
# voyageai>=0.2.0
# cohere>=5.0.0